+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove unnecessary variables in quantity update styles
 - Components structure
 - Add tests in textarea component
+- Remove default slot in input-wrapper component
 
 ### Fixed
 - Typo in Message.selectors.json file